pyproject.toml
[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = "deployments/local/env_files/test.env"
filterwarnings = "ignore::DeprecationWarning"

[tool.ruff]
line-length = 100
select = ["ALL"]
fix = true

ignore = [
    "D100", # docstrings
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",

    "ANN001",
    "ANN003",
    "ANN101", # annotations for self
    "ANN102", # annotations for cls
    "ANN201",
    "ANN202",

    "PLR0913", # too many arguments to function call
    "TD003", # issue link in TODO
    "FIX002", # to do todo :p
    "A002",
    "A003", # attribute "id" shadows builtin
    "SLF001", # private member accessed
    "FA102", # we don't need support for older python versions
    "DJ001",
    "RUF012",
    "ISC001",
    "PLR",
    "FBT002",
    "ARG002",
    "G004",
    "C901",
]

extend-exclude = ["migrations/*", "*_test.py", "conftest.py"]

[tool.mypy]
exclude = ["migrations/*", "*_test.py", "^conftest.py&"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "3.13.*"
pydantic = "2.9.*"
pydantic-settings = "2.6.*"
sqlalchemy = {version = "2.0.*", extras = ["asyncio"]}
asyncpg = "0.30.*"
alembic = "1.13.*"
pytz = "2024.2"
numexpr = "2.10.*"
python-dateutil = "2.9.*"
celery = "5.4.*"
pytest = "8.3.*"
pytest-mock = "3.14.*"
pytest-dotenv = "0.5.*"
pytest-asyncio = "0.24.*"
testcontainers = {version = "4.8.*", extras = ["postgres"]}
pre-commit = "4.0.*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
